Separated Grid and Codebok config. Updated travis config.
Grid and Codebook confifuration have been separated to independent
structs. This allows for potentially easier testing and refactoring.
Lastly, the config item names are more obvious when separated. These
change should make it easier to remove data argument as part of creating
a new map, which does not seem to be semantically correct as data plays
major role merely in actual codebook init: random or PCA.