Updated gonum dependency. Travis tests for go 1.9
gonum has moved to a new Go package repository. This commits updates all
the package imports to gonum.org/v1/gonum/mat and changes few source
files to to this change.

Another improvement this commit brings is dependency management. I decided to
use go dep tool which will become a part of go toolchain and having had a
positive experience with it in past I chose to use it for this project.

Finally TravisCI tests now cover go 1.8 and 1.9