Fix batch training
BMU neighbourhood vector cache had incorrect size allocated. Instead of
allocating it based on the input vector dimension we need to
pre-allocate it to the number of model vectors which is the length or
width of the BMU distance matrix.

If the dimension of the input vector is lower than the number of units
we could be accessing invalid memory space triggering a panic. See line
303 and 310.